+++ R/name_backbone.r
@@ -27,8 +27,8 @@
 #' (optional)
 #' @param strict (logical) If `TRUE` it (fuzzy) matches only the given name,
 #' but never a taxon in the upper classification (optional)
-#' @param verbose (logical) should the function give back more results.
-#' See function `name_backbone_verbose()`
+#' @param verbose (logical) should the function give back more (less reliable) 
+#' results. See function `name_backbone_verbose()`
 #'
 #' @return For `name_backbone`, a data.frame for a single taxon with many
 #' columns. For `name_backbone_verbose`, a larger number of results in a 
@@ -73,12 +73,25 @@ name_backbone <- function(name, rank=NULL, kingdom=NULL, phylum=NULL,
          class=class, order=order, family=family, genus=genus,
          strict=as_log(strict), verbose = verbose, offset=start, limit=limit))
   tt <- gbif_GET(url, args, FALSE, curlopts)
-  input_args <- args[!names(args) %in% c("strict","verbose","start","limit","curlopts")]
-  input_args <- stats::setNames(input_args,paste0("verbatim_",names(input_args)))
-  tt <- c(tt,input_args)
-  out <- tibble::as_tibble(tt[!names(tt) %in% c("alternatives", "note")])
+    input_args_clean <- args[!names(args) %in% c("strict","verbose","start","limit","curlopts")]
+    input_args_clean <- stats::setNames(input_args_clean,paste0("verbatim_",names(input_args_clean)))
+    tt <- c(tt,input_args_clean)
+  if(verbose) {
+    alternatives <- tt[["alternatives"]]
+    alternatives <- lapply(alternatives,function(x) c(x,input_args_clean))
+    alternatives <- bind_rows(lapply(alternatives,tibble::as_tibble))
+    accepted <- tibble::as_tibble(tt)
+    out <- bind_rows(list(accepted,alternatives))
+    out <- out[!colnames(out) %in% c("alternatives", "note")]
+  } else {
+    out <- tibble::as_tibble(tt[!names(tt) %in% c("alternatives", "note")])
+  }
+  col_idx <- grep("verbatim_", names(out))
+  ordering <- c((1:ncol(out))[-col_idx],col_idx)
+  out <- unique(out[, ordering])
   structure(out, args = args, note = tt$note, type = "single")
 }
+  
 
 #' @export
 #' @rdname name_backbone