+++ R/zzz.r
@@ -707,3 +707,51 @@ fail_with <- function(default = NULL, f, quiet = FALSE) {
 }
 
 cat_n <- function(..., sep = "") cat(..., "\n", sep = sep)
+
+pc <- function(..., collapse = "") paste0(..., collapse = collapse)
+
+check_user <- function(x) {
+  z <- if (is.null(x)) Sys.getenv("GBIF_USER", "") else x
+  if (z == "") getOption("gbif_user", stop("supply a username")) else z
+}
+
+check_pwd <- function(x) {
+  z <- if (is.null(x)) Sys.getenv("GBIF_PWD", "") else x
+  if (z == "") getOption("gbif_pwd", stop("supply a password")) else z
+}
+
+check_email <- function(x) {
+  z <- if (is.null(x)) Sys.getenv("GBIF_EMAIL", "") else x
+  if (z == "") getOption("gbif_email", stop("supply an email address")) else z
+}
+
+check_inputs <- function(x) {
+  if (is.character(x)) {
+    # replace newlines
+    x <- gsub("\n|\r|\\s+", "", x)
+    # validate
+    tmp <- jsonlite::validate(x)
+    if (!tmp) stop(attr(tmp, "err"))
+    x
+  } else {
+    jsonlite::toJSON(x)
+  }
+}
+
+is_download_key <- function(x) grepl("^[0-9]{7}-[0-9]{15}$",x)
+
+bind_rows <- function(x) tibble::as_tibble(data.table::rbindlist(x,fill=TRUE))
+
+getsize <- function(x) {
+  round(x/10 ^ 6, 2)
+}
+
+prep_output <- function(x) {
+  list(
+    meta = data.frame(offset = x$offset, limit = x$limit,
+                      endofrecords = x$endOfRecords, count = x$count,
+                      stringsAsFactors = FALSE),
+    results = tibble::as_tibble(x$results)
+  )
+}
+